requirements.txt
supabase>=2
python-dotenv>=1.0